Compute Weighted TF-coactivity Network Analysis (WTCNA)

Construct a weighted signed or unsigned network using TF activity to cluster protein regulators into modules that share similar activity patterns. Each TF module will have a sample-level score represented by the eigenvalue of the module.

Usage

compute.WTCNA(
  TFs.matrix,
  network.type = "signed",
  clustering.method = "ward.D2",
  minMod = 15,
  corr_mod = 0.9,
  cor_type = "p",
  return = T
)

Arguments

TFs.matrix

Matrix of TF activity (samples x TFs).

network.type

Network type: "signed", "unsigned", "signed hybrid", or "distance". Default is "signed".

clustering.method

Clustering method for hierarchical clustering. Default is "ward.D2".

minMod

Minimum number of TFs per module. Default is 15.

corr_mod

Correlation threshold (0–1) for merging similar modules. Default is 0.9.

cor_type

Correlation type for adjacency calculation: "p" (Pearson), "s" (Spearman). Default is "p".

return

Logical, whether to save output plots and module list to "Results/". Default is TRUE.

Value

A named list with:

TFs module matrix: Matrix of module eigengenes (samples x modules).

TFs colors: Vector of module colors assigned to each TF.

TFs per module: List of TF names in each module.

Proportion of variance: Matrix of variance explained per module.

References

Langfelder, P., & Horvath, S. (2008). WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics, 9, 559. https://doi.org/10.1186/1471-2105-9-559

Examples

data("tfs")
network <- compute.WTCNA(tfs, corr_mod = 0.9, clustering.method = "ward.D2", return = FALSE)
#> Creating weighted TF-coactivity network......................................................................
#> 
#> Warning: executing %dopar% sequentially: no parallel backend registered
#>    Power SFT.R.sq slope truncated.R.sq mean.k. median.k. max.k.
#> 1      1    0.351  5.37          0.948 398.000  401.0000 448.00
#> 2      2    0.325 -1.98          0.891 216.000  214.0000 286.00
#> 3      3    0.741 -2.74          0.936 123.000  117.0000 195.00
#> 4      4    0.866 -2.55          0.964  72.500   66.3000 140.00
#> 5      5    0.918 -2.36          0.972  44.500   38.6000 104.00
#> 6      6    0.940 -2.13          0.977  28.300   22.8000  79.90
#> 7      7    0.948 -1.97          0.979  18.600   14.1000  62.60
#> 8      8    0.958 -1.81          0.982  12.600    8.8900  49.90
#> 9      9    0.971 -1.70          0.993   8.770    5.6700  40.40
#> 10    10    0.973 -1.62          0.987   6.280    3.6700  33.20
#> 11    12    0.967 -1.55          0.979   3.450    1.6100  24.00
#> 12    13    0.969 -1.51          0.982   2.640    1.0800  20.70
#> 13    14    0.975 -1.48          0.988   2.050    0.7310  18.00
#> 14    15    0.944 -1.48          0.957   1.620    0.5080  15.70
#> 15    16    0.953 -1.45          0.971   1.300    0.3540  13.90
#> 16    17    0.958 -1.42          0.977   1.060    0.2520  12.30
#> 17    18    0.954 -1.40          0.970   0.869    0.1770  10.90
#> 18    19    0.966 -1.37          0.981   0.722    0.1260   9.78
#> 19    20    0.964 -1.36          0.974   0.606    0.0925   8.78
#> Choosing 5 as soft-threshold......................................................................
#> 
#> Calculating nodes adjacency and topological overlapping nodes.................................................
#> 
#> ..connectivity..
#> ..matrix multiplication (system BLAS)..
#> ..normalization..
#> ..done.
#>  ..cutHeight not given, setting it to 28.3  ===>  99% of the (truncated) height range in dendro.
#>  ..done.
#> Calculating eigenvectors from modules.................................................
#> 
#> [1] "Merging modules significantly correlated with 0.9........"